+++ pkg/web/src/components/atoms/ImageIcon.tsx
@@ -1,37 +1,37 @@
 import styled from 'styled-components'
 export const ImageIcon = styled.i`
   & {
-    box-sizing: border-box;
     position: relative;
+    box-sizing: border-box;
     display: block;
-    transform: scale(var(--ggs, 1));
     width: 16px;
     height: 12px;
     overflow: hidden;
-    box-shadow: 0 0 0 2px;
     border-radius: 2px;
+    box-shadow: 0 0 0 2px;
+    transform: scale(var(--ggs, 1));
   }
   &::after,
   &::before {
-    content: '';
-    display: block;
-    box-sizing: border-box;
     position: absolute;
+    box-sizing: border-box;
+    display: block;
+    content: '';
     border: 2px solid;
   }
   &::after {
-    transform: rotate(45deg);
-    border-radius: 3px;
-    width: 14px;
-    height: 14px;
     top: 8px;
     left: 5px;
+    width: 14px;
+    height: 14px;
+    border-radius: 3px;
+    transform: rotate(45deg);
   }
   &::before {
+    top: 2px;
+    left: 2px;
     width: 6px;
     height: 6px;
     border-radius: 100%;
-    top: 2px;
-    left: 2px;
   }
 `
