+++ pkg/web/src/pages/browser/[...pathes]/components/ProjectNameInput.tsx
@@ -8,9 +8,9 @@ import styled from 'styled-components'
 
 const InputFormProject = styled.form`
   display: flex;
-  justify-content: center;
-  align-items: center;
   flex-direction: column;
+  align-items: center;
+  justify-content: center;
 `
 
 export const ProjectNameInput = (props: { onComplete?: () => void }) => {
